vendorMS/requirements.txt
Django
djangorestframework
python-dotenv
APScheduler